+++ test/must.go
@@ -1,10 +1,10 @@
 package test
 
-func MustBytes(bytes []byte, err error) []byte {
+func MustBytes(bites []byte, err error) []byte {
 	if err != nil {
 		panic(err)
 	}
-	return bytes
+	return bites
 }
 
 func MustString(str string, err error) string {